Criando Entidade AuditLog e adicionando verificação do usuario criador para manipular Produto.

+++ src/main/java/br/com/indra/roger_willians/service/ProdutoService.java
@@ -3,9 +3,12 @@ package br.com.indra.roger_willians.service;
 import br.com.indra.roger_willians.exception.RecursoNaoEncontradoException;
 import br.com.indra.roger_willians.model.Categoria;
 import br.com.indra.roger_willians.model.Produto;
+import br.com.indra.roger_willians.model.enums.AcaoAuditoria;
+import br.com.indra.roger_willians.model.enums.TipoEntidade;
 import br.com.indra.roger_willians.repository.ProdutoRepository;
 import br.com.indra.roger_willians.service.dto.ProdutoDTO;
 import br.com.indra.roger_willians.service.dto.ProdutoResponseDTO;
+import br.com.indra.roger_willians.service.dto.ProdutoSnapShotDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -21,6 +24,7 @@ public class ProdutoService {
     private final ProdutoRepository produtoRepository;
     private final HistoricoPrecoService historicoPrecoService;
     private final CategoriaService categoriaService;
+    private final AuditLogService auditLogService;
 
     public List<ProdutoResponseDTO> findAll(){
 
@@ -113,7 +117,7 @@ public class ProdutoService {
     }
 
     @Transactional
-    public ProdutoResponseDTO cadastrarProduto(ProdutoDTO dto) {
+    public ProdutoResponseDTO cadastrarProduto(ProdutoDTO dto, UUID usuarioId) {
 
         if (produtoRepository.findBySku(dto.sku()).isPresent()) {
             throw new IllegalArgumentException("Já existe um produto cadastrado com o SKU: " + dto.sku());
@@ -121,65 +125,96 @@ public class ProdutoService {
 
         Categoria categoria = categoriaService.buscarEntidadePorId(dto.categoriaId());
 
-        Produto produtoNovo = converterParaEntidade(dto, categoria);
+        Produto produtoSalvo = produtoRepository.save(converterParaEntidade(dto, categoria, usuarioId));
 
-        Produto produtoSalvo = produtoRepository.save(produtoNovo);
+
+        ProdutoSnapShotDTO produtoCriado = snapShot(produtoSalvo);
+
+        auditLogService.registrarLog(
+                TipoEntidade.PRODUTO,
+                produtoSalvo.getId(),
+                AcaoAuditoria.CRIACAO,
+                null,
+                produtoCriado,
+                usuarioId
+        );
 
         return converterParaDTO(produtoSalvo);
     }
 
     @Transactional
-    public ProdutoResponseDTO atualizarProduto(UUID id, ProdutoDTO dto) {
+    public ProdutoResponseDTO atualizarProduto(UUID id, UUID usuarioId, ProdutoDTO dto) {
         final var produtoAtual = produtoRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Produto não encontrado"));
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Produto não encontrado"));
 
         if (!produtoAtual.getSku().equals(dto.sku()) &&
                 produtoRepository.findBySku(dto.sku()).isPresent()) {
             throw new IllegalArgumentException("Já existe outro produto cadastrado com o SKU: " + dto.sku());
         }
+        if(!produtoAtual.getVendedorId().equals(usuarioId)){
+            throw new IllegalArgumentException("Apenas o vendedor pode atualizar o produto.");
+        }
+        ProdutoSnapShotDTO produtoAntigo = snapShot(produtoAtual);
 
-        Categoria categoria = categoriaService.buscarEntidadePorId(dto.categoriaId());
-
-        atualizarDadosEntidade(produtoAtual, dto, categoria);
+        atualizarDadosEntidade(produtoAtual, dto, categoriaService.buscarEntidadePorId(dto.categoriaId()));
 
         Produto produtoAtualizado = produtoRepository.save(produtoAtual);
+
+        auditLogService.registrarLog(
+                TipoEntidade.PRODUTO,
+                produtoAtualizado.getId(),
+                AcaoAuditoria.ATUALIZACAO,
+                produtoAntigo,
+                snapShot(produtoAtualizado),
+                usuarioId
+        );
 
         return converterParaDTO(produtoAtualizado);
     }
 
-
     @Transactional
-    public ProdutoResponseDTO atualizarPreco(UUID id, BigDecimal precoNovo) {
-        final var produto = produtoRepository.findById(id)
+    public ProdutoResponseDTO atualizarPreco(UUID id, UUID usuarioId, BigDecimal precoNovo) {
+        var produto = produtoRepository.findById(id)
                 .orElseThrow(() -> new RecursoNaoEncontradoException("Produto não encontrado"));
-        Produto produtoAtualizado;
+        if(!produto.getVendedorId().equals(usuarioId)){
+            throw new IllegalArgumentException("Apenas o vendedor pode atualizar o produto.");
+        }
+        var precoAntigo = produto.getPreco();
+        ProdutoSnapShotDTO produtoAntigo = snapShot(produto);
 
-        BigDecimal precoAntigo = produto.getPreco();
-
-        produto.setPreco(precoNovo);
-
-        if (precoAntigo.compareTo(precoNovo) != 0) {
-
+        if (produto.getPreco().compareTo(precoNovo) != 0) {
 
             produto.setPreco(precoNovo);
-            produtoAtualizado = produtoRepository.save(produto);
+            produtoRepository.save(produto);
 
-
-            historicoPrecoService.registrarHistorico(produtoAtualizado, precoAntigo, precoNovo);
+            historicoPrecoService.registrarHistorico(produto, precoAntigo, precoNovo);
         } else{
             throw new IllegalArgumentException("O preço novo é igual ao preço atual.");
         }
 
+        ProdutoSnapShotDTO produtoNovo = snapShot(produto);
 
-        return converterParaDTO(produtoAtualizado);
+        auditLogService.registrarLog(
+                TipoEntidade.PRODUTO,
+                produto.getId(),
+                AcaoAuditoria.ATUALIZACAO,
+                produtoAntigo,
+                produtoNovo,
+                usuarioId
+        );
+
+        return converterParaDTO(produto);
 
     }
 
     @Transactional
-    public ProdutoResponseDTO reativarProduto(UUID id) {
+    public ProdutoResponseDTO reativarProduto(UUID id, UUID usuarioId) {
 
         Produto produto = produtoRepository.findById(id)
                 .orElseThrow(() -> new RecursoNaoEncontradoException("Produto não encontrado"));
+        if(!produto.getVendedorId().equals(usuarioId)){
+            throw new IllegalArgumentException("Apenas o vendedor pode atualizar o produto.");
+        }
 
         if(produto.getAtivo()) {
             throw new IllegalArgumentException("O produto já está ativo.");
@@ -190,20 +225,35 @@ public class ProdutoService {
         return converterParaDTO(produtoRepository.save(produto));
     }
 
-
     @Transactional
-    public void deletarProduto(UUID id) {
+    public void deletarProduto(UUID id, UUID usuarioId) {
         final var produto = produtoRepository.findById(id)
                 .orElseThrow(() -> new RecursoNaoEncontradoException("Produto não encontrado"));
-
+        if(!produto.getVendedorId().equals(usuarioId)){
+            throw new IllegalArgumentException("Apenas o vendedor pode deletar o produto.");
+        }
+        ProdutoSnapShotDTO produtodeletado = snapShot(produto);
         produtoRepository.delete(produto);
+
+        auditLogService.registrarLog(
+                TipoEntidade.PRODUTO,
+                produto.getId(),
+                AcaoAuditoria.EXCLUSAO,
+                produtodeletado,
+                null,
+                usuarioId
+        );
+
     }
 
     @Transactional
-    public void inativarProduto(UUID id) {
+    public void inativarProduto(UUID id, UUID usuarioId) {
 
         Produto produto = produtoRepository.findById(id)
                 .orElseThrow(() -> new RecursoNaoEncontradoException("Produto não encontrado"));
+        if(!produto.getVendedorId().equals(usuarioId)){
+            throw new IllegalArgumentException("Apenas o vendedor pode atualizar o produto.");
+        }
 
         produto.setAtivo(false);
 
@@ -227,11 +277,12 @@ public class ProdutoService {
                 produto.getSku(),
                 produto.getPreco(),
                 produto.getQuantidadeEstoque(),
-                produto.getNotaMedia()
+                produto.getNotaMedia(),
+                produto.getVendedorId()
         );
     }
 
-    private Produto converterParaEntidade(ProdutoDTO dto, Categoria categoria) {
+    private Produto converterParaEntidade(ProdutoDTO dto, Categoria categoria, UUID usuarioId) {
         Produto produto = new Produto();
         produto.setNome(dto.nome());
         produto.setDescricao(dto.descricao());
@@ -239,6 +290,7 @@ public class ProdutoService {
         produto.setPreco(dto.preco());
         produto.setPrecoCusto(dto.precoCusto());
         produto.setQuantidadeEstoque(dto.quantidadeEstoque());
+        produto.setVendedorId(usuarioId);
 
         produto.setCategoria(categoria);
 
@@ -254,5 +306,21 @@ public class ProdutoService {
         produtoExistente.setQuantidadeEstoque(dto.quantidadeEstoque());
 
         produtoExistente.setCategoria(categoria);
+    }
+
+    private ProdutoSnapShotDTO snapShot(Produto produto) {
+        return new ProdutoSnapShotDTO(
+                produto.getId(),
+                produto.getNome(),
+                produto.getDescricao(),
+                produto.getSku(),
+                produto.getPreco(),
+                produto.getPrecoCusto(),
+                produto.getQuantidadeEstoque(),
+                produto.getCategoria().getId(),
+                produto.getNotaMedia(),
+                produto.getVendedorId(),
+                produto.getAtivo()
+        );
     }
 }
